Rename math.Uint256le to ecmath.Scalar for brevity
