+++ crypto/ed25519/ecmath/point_test.go
@@ -1,4 +1,4 @@
-package math
+package ecmath
 
 import "testing"
